created LoadingSpinner component
