+++ src/App.js
@@ -1,11 +1,11 @@
 import React, {useEffect, useState} from "react";
 import Grid from "@material-ui/core/Grid";
-import CircularProgress from '@material-ui/core/CircularProgress';
 
 import axios from "./axiosInstance";
-import URLInput from "./components/URLInput"
+import SearchBar from "./components/SearchBar"
 import Graph from "./components/Graph";
 import Screenshot from "./components/Screenshot";
+import LoadingSpinner from "./components/LoadingSpinner";
 import useKeyPress from "./hooks/useKeyPress"
 import traverseToNewNode from "./utils/graph/traversal";
 import getNodeAtPoint from "./utils/graph/node_at_point";
@@ -88,7 +88,7 @@ function App() {
         >
 
             <Grid item xs={12} style={{height: "5vh"}}>
-                <URLInput getWebPage={getWebPage}/>
+                <SearchBar getWebPage={getWebPage}/>
             </Grid>
 
             {
@@ -114,9 +114,7 @@ function App() {
                     : loading
                     ? (
                         <Grid>
-                            <div style={{display: "flex", justifyContent: "center", width: "100vw"}}>
-                                <CircularProgress/>
-                            </div>
+                            <LoadingSpinner/>
                         </Grid>
                     )
                     : null
