+++ src/components/LoadingSpinner/loading_spinner.module.css
@@ -1,0 +1,5 @@
+.loadingSpinnerContainer {
+    display: flex;
+    justify-content: center;
+    width: 100vw;
+}
